[FIX] Cash Deposit Match  (#83)
Merging - approved by Vyas previously and given verbal go-ahead to merge as the previously approval disappeared after I merged in changes from two other branches. tested locally.

+++ apps/backend/src/location/entities/master-location-data.entity.ts
@@ -14,7 +14,7 @@ export class LocationEntity {
   @Column('varchar', { length: 15, nullable: false })
   method: string;
 
-  @Column({ type: 'numeric', nullable: false })
+  @Column({ type: 'int4', nullable: false })
   pt_location_id: number;
 
   @Column('varchar', { length: 255, nullable: false })
